Module render_matic::component

The "component" part of the entity-component model.

Components are collections of data providing all the information a system needs to execute a desired behavior. For example, in order to draw a tree to the screen, the DrawingSystem needs to know how big the tree should be, where it should be drawn, what texture or part of a texture contains the picture we want to use, and so on. All of this information is bundled together into a Component with the and placed in the ComponentManager. When it comes time to draw a frame to the screen, the DrawingSystem will ask the ComponentManager for every Component of each drawable type and use that information to draw all of the currently existing drawable components to the screen.

Structs

Component
A single data object with no associated behavior.

ComponentManager
The owner for all Components. Controls creation, access, and deletion.

Enums

ComponentData
Stores a Component's data. Use match or if let constructs to access members of a particular ComponentData variant.

ComponentType
An enum for the different types of Components. Rust provides std::mem::discriminant for looking at which variant an enum object belongs to, and it works for enum variants that have attached data. However, it's a little clunky and it's simpler to use a plain enum for variant identification and comparison.